Updates
"Major Change": "Had to make the hide button not reload all of the messages because of too many api calls.",

"HiddenPeople": "The hidden people now caches.",

"OneDrive Support": "You can now set up your onedrive to have it on multipule computers and it will remember who you hid. You have to make a folder called BetterDiscordConfigs on there, if it finds a file it will use it instead if the file in your plugins folder. The file should be called HidesPeopleOnYourList.config.json",

"Upcomming OneDrive Support toggle": "It is to be implimented in settings not gotten around to it yet though."